feat: order export

+++ src/CAServer.Application.Contracts/Commons/StringHelper.cs
@@ -36,7 +36,7 @@ public static class StringHelper
         return decimal.TryParse(s, NumberStyles.Float, CultureInfo.InvariantCulture, out var result) ? result : defaultValue;
     }
     
-    public static double SafeToInt(this string s, int defaultValue = 0)
+    public static int SafeToInt(this string s, int defaultValue = 0)
     {
         return int.TryParse(s, NumberStyles.Float, CultureInfo.InvariantCulture, out var result) ? result : defaultValue;
     }
